+++ build.clj
@@ -1,3 +1,7 @@
+; This Source Code Form is subject to the terms of the Mozilla Public
+; License, v. 2.0. If a copy of the MPL was not distributed with this
+; file, You can obtain one at https://mozilla.org/MPL/2.0/.
+
 (ns build
   (:require
    [clojure.java.io :as io]
@@ -11,20 +15,21 @@
 (defn make-opts [opts]
   (let [version (make-version)
         lib 'io.github.noahtheduke/fluent-clj]
-    (assoc opts
-      :lib lib
-      :version version
-      :scm {:url "https://github.com/noahtheduke/fluent-clj"
-            :tag (str "v" version)}
-      :main 'fluent-clj.main
-      :basis (b/create-basis {:project "deps.edn"})
-      :class-dir "target/classes"
-      :jar-file (str (io/file "target" (format "%s-%s.jar" (name lib) version)))
-      :uber-file (str (io/file "target" (format "%s-%s-standalone.jar" (name lib) version)))
-      :src-dirs ["src"]
-      :resource-dirs ["resources"]
-      :pom-data [[:licenses
-                  [:license [:name "MPL-2.0"] [:url "https://mozilla.org/MPL/2.0"]]]])))
+    (merge {:lib lib
+            :version version
+            :scm {:url "https://github.com/noahtheduke/fluent-clj"
+                  :tag (str "v" version)}
+            :main 'fluent-clj.main
+            :basis (b/create-basis {:project "deps.edn"})
+            :class-dir "target/classes"
+            :jar-file (str (io/file "target" (format "%s-%s.jar" (name lib) version)))
+            :uber-file (str (io/file "target" (format "%s-%s-standalone.jar" (name lib) version)))
+            :src-dirs ["src"]
+            :java-src-dirs ["src"]
+            :resource-dirs ["resources"]
+            :pom-data [[:licenses
+                        [:license [:name "MPL-2.0"] [:url "https://mozilla.org/MPL/2.0"]]]]}
+           opts)))
 
 (defn clean [opts]
   (let [opts (make-opts opts)]
@@ -35,15 +40,30 @@
     (b/copy-dir {:src-dirs (concat (:src-dirs opts) (:resource-dirs opts))
                  :target-dir (:class-dir opts)})))
 
+(defn javac [opts]
+  (let [opts (make-opts opts)]
+    (println "Compiling" (str/join ", " (:java-src-dirs opts)))
+    (b/javac {:src-dirs (:java-src-dirs opts)
+              :class-dir (:class-dir opts)
+              :basis (:basis opts)
+              :javac-opts (:javac-opts opts)})))
+
 (defn jar [opts]
   (let [opts (make-opts opts)]
     (copy-src opts)
+    (javac opts)
     (b/jar opts)
     (println "Created" (str (b/resolve-path (:jar-file opts))))))
+
+(defn write-pom
+  [opts]
+  (let [opts (make-opts opts)]
+    (b/write-pom opts)))
 
 (defn uberjar [opts]
   (let [opts (make-opts opts)]
     (copy-src opts)
+    (javac opts)
     (b/write-pom opts)
     (b/compile-clj opts)
     (b/uber opts)
@@ -52,8 +72,10 @@
 (defn deploy [opts]
   (let [opts (make-opts opts)]
     (clean opts)
+    (copy-src opts)
+    (javac opts)
     (b/write-pom opts)
-    (jar opts)
+    (b/jar opts)
     (dd/deploy {:installer :remote
                 :artifact (b/resolve-path (:jar-file opts))
                 :pom-file (b/pom-path opts)})))